+++ xss-core/src/main/java/org/stage/xss/core/config/XssStageConfigurer.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.stage.xss.core.XssFilterAop;
+import org.stage.xss.core.engine.aop.XssFilterAopEngine;
 import org.stage.xss.core.spi.XssFilter;
 
 @Configuration
@@ -14,8 +14,8 @@ public class XssStageConfigurer{
     List<XssFilter> xssFilterList;
 
     @Bean
-    public XssFilterAop xssFilterAop(){
-        return new XssFilterAop(xssFilterList);
+    public XssFilterAopEngine xssFilterAop(){
+        return new XssFilterAopEngine(xssFilterList);
     }
 
 }
